requirements.txt
streamlit>=1.35.0
gitpython>=3.1.43
pyyaml>=6.0.1
pydantic>=2.7.0
